content-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
python-telegram-bot==20.7
replicate==0.22.0
apscheduler==3.10.4
beautifulsoup4==4.12.2
crewai
langchain
google-generativeai==0.3.2
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
lxml==4.9.3